# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  dto.ArticleResponse:
    properties:
      category:
        $ref: '#/definitions/dto.CategoryResponse'
      content:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      id:
        type: string
      imageUrl:
        type: string
      slug:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  dto.CategoryResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
      updatedAt:
        type: string
    type: object
  dto.CreateArticleRequest:
    properties:
      categoryId:
        type: string
      content:
        minLength: 10
        type: string
      title:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - categoryId
    - content
    - title
    type: object
...
  dto.Meta:
    properties:
      currentPage:
        type: integer
      limit:
        type: integer
      totalPages:
        type: integer
      totalRecords:
        type: integer
    type: object
  dto.PaginationResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/dto.Meta'
    type: object
...
  dto.UpdateArticleRequest:
    properties:
      categoryId:
        type: string
      content:
        minLength: 10
        type: string
      title:
        maxLength: 255
        minLength: 3
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: Backend API for SmartFarming fullstack application using Gin, GORM,
    PostgreSQL, MongoDB, and Redis.
  title: SmartFarming API
  version: "1.0"
paths:
  /articles:
    get:
      consumes:
      - application/json
      description: Lists articles with pagination, search by title/content, and filter
        by categoryId
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit per page
        in: query
        name: limit
        type: integer
      - description: Search by title or content
        in: query
        name: search
        type: string
      - description: Filter by Category UUID
        in: query
        name: categoryId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: List Articles with search, pagination and filtering
      tags:
      - Articles
    post:
      consumes:
      - application/json
      description: Creates a new article (Admin/Operator only)
      parameters:
      - description: Create Article request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateArticleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ArticleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Create an Article
      tags:
      - Articles
  /articles/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an article by its UUID (Admin/Operator only)
      parameters:
      - description: Article UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete Article
      tags:
      - Articles
    get:
      consumes:
      - application/json
      description: Fetches details of an article by its UUID or Slug
      parameters:
      - description: Article UUID or Slug
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ArticleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Get Article by ID or Slug
      tags:
      - Articles
    put:
      consumes:
      - application/json
      description: Updates article details by its UUID (Admin/Operator only)
      parameters:
      - description: Article UUID
        in: path
        name: id
        required: true
        type: string
      - description: Update Article request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateArticleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ArticleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Update Article
      tags:
      - Articles
  /articles/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Uploads an image asset for a specific article (Admin/Operator only)
      parameters:
      - description: Article UUID
        in: path
        name: id
        required: true
        type: string
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ArticleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Upload Article Image
      tags:
      - Articles
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your JWT token to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"